moved compileAndUpsertPlace() to modules.js
